Removed wmin executable
